surpress warnings

+++ libmp3lame/tags.c
@@ -521,7 +521,8 @@ write_frame(unsigned char *p, frame_struct *frame)
         }
     }
     if (frame->length != p - p0) {
-        fprintf(stderr, "Internal error: %ld != %ld\n", frame->length, p - p0);
+        fprintf(stderr, "Internal error: %d != %d\n",
+		(int)frame->length, (int)(p - p0));
         exit(1);
     }
     return p;
